Refactor AppElement using jsx
